Move 4C related stuff out of function
